# sessionctx/stmtctx: do not use copy-on-read for GetWarnings function