<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: close #39702

Problem Summary:

### What is changed and how it works?

7 min 53s => 9.32s
(B.T.W, if we don't add error stack on decimal truncate error, the execution time can furthur reduce to 5s)

![image](https://user-images.githubusercontent.com/1420062/206188499-a9d9f54d-3d6d-497e-92e0-9688c3a443f5.png)


Before the change:

```
mysql> insert into t1 (c1) select * from t2;
Query OK, 491520 rows affected, 65535 warnings (7 min 53.87 sec)
Records: 491520  Duplicates: 0  Warnings: 65535
```

After:

```
mysql> insert into t1 (c1) select * from t2;
Query OK, 491520 rows affected, 65535 warnings (9.32 sec)
Records: 491520  Duplicates: 0  Warnings: 65535
```

### Check List

Tests <!-- At least one of them must be included. -->

- [ ] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)

See https://github.com/pingcap/tidb/issues/39702 reproduce process.

- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```